# gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: Erabooru API
    version: 0.0.1
servers:
    - url: http://localhost:8080
paths:
    /api/admin/regenerate:
        post:
            tags:
                - admin
            summary: regenerate admin
            description: Regenerate search index and metadata
            operationId: admin#regenerate
            responses:
                "204":
                    description: No Content response.
    /api/media:
        get:
            tags:
                - media
            summary: list media
            description: List media items
            operationId: media#list
            parameters:
                - name: q
                  in: query
                  allowEmptyValue: true
                  schema:
                    type: string
                    example: Maxime quod autem ducimus.
                  example: Optio vel officiis dolore.
                - name: page
                  in: query
                  allowEmptyValue: true
                  schema:
                    type: integer
                    example: 2143469068996669834
                    format: int64
                  example: 2791063455337779646
                - name: page_size
                  in: query
                  allowEmptyValue: true
                  schema:
                    type: integer
                    example: 5467623489395213036
                    format: int64
                  example: 3538227376797016209
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/MediaList'
                            example:
                                media:
                                    - format: Ea voluptate qui delectus rerum.
                                      height: 4478699467164243213
                                      id: Iusto culpa ex repellat est maiores.
                                      url: Laboriosam aut recusandae.
                                      width: 4269257746853478508
                                    - format: Ea voluptate qui delectus rerum.
                                      height: 4478699467164243213
                                      id: Iusto culpa ex repellat est maiores.
                                      url: Laboriosam aut recusandae.
                                      width: 4269257746853478508
                                    - format: Ea voluptate qui delectus rerum.
                                      height: 4478699467164243213
                                      id: Iusto culpa ex repellat est maiores.
                                      url: Laboriosam aut recusandae.
                                      width: 4269257746853478508
                                    - format: Ea voluptate qui delectus rerum.
                                      height: 4478699467164243213
                                      id: Iusto culpa ex repellat est maiores.
                                      url: Laboriosam aut recusandae.
                                      width: 4269257746853478508
                                total: 1881487347525540260
    /api/media/{id}:
        delete:
            tags:
                - media
            summary: delete media
            description: Delete media item
            operationId: media#delete
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                    example: Quis sed quisquam minus nulla est.
                  example: Iusto corrupti recusandae fugit iste distinctio voluptatibus.
            responses:
                "204":
                    description: No Content response.
        get:
            tags:
                - media
            summary: get media
            description: Get media item by ID
            operationId: media#get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                    example: Cupiditate ut et.
                  example: Facere rerum quia autem.
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/MediaDetail'
                            example:
                                duration: 6666198724408845616
                                format: Molestias earum est consequatur aperiam deserunt.
                                height: 3667668171383177776
                                id: Possimus ullam illo qui facere.
                                size: 6998739584799792186
                                tags:
                                    - Architecto quod repellat harum fugit.
                                    - Minima esse voluptatem id.
                                url: Possimus est est assumenda sint alias.
                                width: 8895310890499510412
    /api/media/{id}/tags:
        post:
            tags:
                - media
            summary: updateTags media
            description: Update tags for media item
            operationId: media#updateTags
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                    example: Reiciendis laborum.
                  example: Iusto autem officia aut voluptatum omnis.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                tags:
                                    type: array
                                    items:
                                        type: string
                                        example: Perferendis soluta repudiandae sapiente.
                                    example:
                                        - Nam quis est necessitatibus omnis accusamus aspernatur.
                                        - Inventore voluptatibus et itaque cupiditate est aperiam.
                                        - Veritatis earum nisi labore.
                                        - Facere sit sint vitae soluta dolorem.
                            example:
                                tags:
                                    - Consequatur rerum qui quidem.
                                    - Et perferendis vel.
                                    - Sit quia.
                                    - Amet et laborum qui neque delectus.
                        example:
                            tags:
                                - Sunt quia magni.
                                - Odit aspernatur eius ut.
                                - Architecto nisi impedit impedit autem.
            responses:
                "204":
                    description: No Content response.
    /api/media/previews:
        get:
            tags:
                - media
            summary: previews media
            description: List media previews
            operationId: media#previews
            parameters:
                - name: q
                  in: query
                  allowEmptyValue: true
                  schema:
                    type: string
                    example: Quis rerum cupiditate a omnis.
                  example: Quod voluptas sed ullam.
                - name: page
                  in: query
                  allowEmptyValue: true
                  schema:
                    type: integer
                    example: 1091163213526196693
                    format: int64
                  example: 4804694174010542949
                - name: page_size
                  in: query
                  allowEmptyValue: true
                  schema:
                    type: integer
                    example: 4321570254515286637
                    format: int64
                  example: 757983125018217480
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/MediaList'
                            example:
                                media:
                                    - format: Ea voluptate qui delectus rerum.
                                      height: 4478699467164243213
                                      id: Iusto culpa ex repellat est maiores.
                                      url: Laboriosam aut recusandae.
                                      width: 4269257746853478508
                                    - format: Ea voluptate qui delectus rerum.
                                      height: 4478699467164243213
                                      id: Iusto culpa ex repellat est maiores.
                                      url: Laboriosam aut recusandae.
                                      width: 4269257746853478508
                                total: 7369881979245154909
    /api/media/upload-url:
        post:
            tags:
                - media
            summary: uploadURL media
            description: Get a presigned upload URL
            operationId: media#uploadURL
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                filename:
                                    type: string
                                    example: Omnis saepe repellendus tempora est.
                            example:
                                filename: Omnis esse provident aut.
                        example:
                            filename: Et eius nemo.
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UploadURLResponse'
                            example:
                                object: Dignissimos sed.
                                url: Vero molestias ad.
    /health:
        get:
            tags:
                - health
            summary: check health
            operationId: health#check
            responses:
                "204":
                    description: No Content response.
components:
    schemas:
        MediaDetail:
            type: object
            properties:
                duration:
                    type: integer
                    description: Video duration in seconds
                    example: 6601154237725332209
                    format: int64
                format:
                    type: string
                    example: Necessitatibus est.
                height:
                    type: integer
                    example: 7804486503803889128
                    format: int64
                id:
                    type: string
                    example: Alias ab et aut quia.
                size:
                    type: integer
                    example: 5490997323526538113
                    format: int64
                tags:
                    type: array
                    items:
                        type: string
                        example: Fugit deserunt occaecati.
                    example:
                        - Id vitae ducimus nam doloremque.
                        - Dolores ipsum in eum aspernatur.
                        - Ut in labore veritatis quam corrupti.
                        - Delectus qui sit fugit sint.
                url:
                    type: string
                    example: Amet quia consequatur dignissimos in.
                width:
                    type: integer
                    example: 3870547163235901958
                    format: int64
            example:
                duration: 7421203309775957695
                format: Et iste expedita optio corrupti assumenda sed.
                height: 5993495206752240411
                id: Non ratione eligendi qui quia doloremque.
                size: 4489052888058278899
                tags:
                    - Nihil quasi rem nulla et sequi aut.
                    - Aut hic voluptas provident sit iure quia.
                    - Quis voluptatem fuga est placeat quia.
                    - Sint vel qui modi similique.
                url: Eligendi architecto quaerat est.
                width: 5481774234563483632
        MediaItem:
            type: object
            properties:
                format:
                    type: string
                    example: Voluptatibus et et.
                height:
                    type: integer
                    example: 3182374220053586339
                    format: int64
                id:
                    type: string
                    example: Porro ad qui voluptates adipisci tenetur.
                url:
                    type: string
                    example: Repellat voluptates qui reprehenderit dolores.
                width:
                    type: integer
                    example: 5085790072468248624
                    format: int64
            example:
                format: Qui nisi dolor est quo ratione.
                height: 681756893210495030
                id: Esse voluptatem dolor in et quae quia.
                url: Tenetur et nostrum atque eaque molestiae nam.
                width: 3578520200240905401
        MediaList:
            type: object
            properties:
                media:
                    type: array
                    items:
                        $ref: '#/components/schemas/MediaItem'
                    example:
                        - format: Mollitia qui officiis consequatur.
                          height: 4064237908605303538
                          id: Quia in quam laborum aliquam sint.
                          url: Expedita iusto explicabo voluptate id.
                          width: 6453942027765999311
                        - format: Mollitia qui officiis consequatur.
                          height: 4064237908605303538
                          id: Quia in quam laborum aliquam sint.
                          url: Expedita iusto explicabo voluptate id.
                          width: 6453942027765999311
                total:
                    type: integer
                    example: 1076505782096345897
                    format: int64
            example:
                media:
                    - format: Mollitia qui officiis consequatur.
                      height: 4064237908605303538
                      id: Quia in quam laborum aliquam sint.
                      url: Expedita iusto explicabo voluptate id.
                      width: 6453942027765999311
                    - format: Mollitia qui officiis consequatur.
                      height: 4064237908605303538
                      id: Quia in quam laborum aliquam sint.
                      url: Expedita iusto explicabo voluptate id.
                      width: 6453942027765999311
                    - format: Mollitia qui officiis consequatur.
                      height: 4064237908605303538
                      id: Quia in quam laborum aliquam sint.
                      url: Expedita iusto explicabo voluptate id.
                      width: 6453942027765999311
                    - format: Mollitia qui officiis consequatur.
                      height: 4064237908605303538
                      id: Quia in quam laborum aliquam sint.
                      url: Expedita iusto explicabo voluptate id.
                      width: 6453942027765999311
                total: 4099752415003345202
            required:
                - media
                - total
        UploadURLResponse:
            type: object
            properties:
                object:
                    type: string
                    example: Voluptatem et minus nisi voluptatem.
                url:
                    type: string
                    example: Aliquid odit in enim iure.
            example:
                object: Sint natus voluptatem consequatur beatae.
                url: Praesentium explicabo.
tags:
    - name: media
      description: Media operations
    - name: admin
      description: Admin operations
    - name: health
      description: Service health check